refactoring chat example
